requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic[email]==2.5.0
google-genai==0.2.2
psycopg2-binary==2.9.9
argon2-cffi==23.1.0
jinja2
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
dnspython==2.4.2
httpx==0.25.1
aiohttp